+++ server/application-server/src/main/java/de/tum/cit/aet/helios/http/HttpClientConfig.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.helios.github;
+package de.tum.cit.aet.helios.http;
 
 import java.io.File;
 import java.util.concurrent.TimeUnit;
@@ -26,8 +26,12 @@ public class HttpClientConfig {
 
   private final Environment environment;
 
-  public HttpClientConfig(Environment environment) {
+  private final RateLimitInfoHolder rateLimitInfoHolder;
+
+
+  public HttpClientConfig(Environment environment, RateLimitInfoHolder rateLimitInfoHolder) {
     this.environment = environment;
+    this.rateLimitInfoHolder = rateLimitInfoHolder;
   }
 
   /**
@@ -49,7 +53,7 @@ public class HttpClientConfig {
     OkHttpClient.Builder builder = new OkHttpClient.Builder();
 
     // Add the rate limit interceptor
-    builder.addInterceptor(new HttpClientRateLimitInterceptor());
+    builder.addInterceptor(new HttpClientRateLimitInterceptor(rateLimitInfoHolder));
 
     if (cacheEnabled) {
       File cacheDir = new File("./build/github-cache");
